fix: kitItem for collectible

+++ database/migrations/2024_11_04_123932_create_collectible_kit_item_table.php
@@ -23,7 +23,7 @@ return new class extends Migration
                 ->constrained()
                 ->cascadeOnDelete()
                 ->cascadeOnUpdate();
-            $table->tinyInteger('condition')->nullable();
+            $table->smallInteger('condition')->nullable();
             $table->timestamps();
         });
     }
